% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Simulated data for tutorial}
\format{
A list of 4 studies with different sample sizes and each of them contains:
\describe{
  \item{X}{The exposure mimics age}
  \item{Y}{The outcome mimics systolic blood pressure}
  \item{M}{The mediators}
}
}
\usage{
data
}
\description{
simulated 4 independent studies for meta-analysis
}
\references{
<https://www.biorxiv.org/content/10.1101/2023.02.06.527391v1.abstract>
}
\keyword{datasets}
