% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueR2Cal.R
\name{TrueR2Cal}
\alias{TrueR2Cal}
\title{TrueR2Cal.R}
\usage{
TrueR2Cal(
  alpha,
  beta,
  r = 3,
  p0,
  p1,
  p2,
  p3,
  CorM,
  corMat = NULL,
  LargeSam = TRUE
)
}
\arguments{
\item{alpha}{The coefficients between the exposure and mediators.}

\item{beta}{The coefficients between mediators and the outcome.}

\item{r}{The direct effect.}

\item{p0}{The number of the true mediators}

\item{p1}{The number of the type I non-mediators.}

\item{p2}{The number of the type II non-mediators.}

\item{p3}{The number of the noise variables.}

\item{CorM}{The correlation indicator (1=independent, 3=correlated).}

\item{corMat}{The correlation matrix of putative mediators (default=NULL).}

\item{LargeSam}{If TRUE, then use large sample to find the true values.}
}
\value{
A list with the elements:
\item{R2.true}{True value of the R2 total mediation effect}
\item{SOS.true}{True value of the Shared Over Simple effect (SOS)}
\item{R2YX.true}{True value of the R2 between X and Y}
\item{R2YM.true}{True value of the R2 between M and Y}
\item{R2YMX.true}{True value of the R2 between Y and XM}
}
\description{
Calculate the true value of R2-based total mediation effect.
}
\examples{
p0 <- 20
p1 <- 20
p2 <- 20
p3 <- 40
alpha <- c(rnorm(p0, mean = 0, sd = 1.5), rep(0, p1), 
rnorm(p2, mean = 0, sd = 1.5), rep(0, p3))
beta <- c(rnorm(p0, mean=0, sd=1.5),
rnorm(p1, mean=0, sd=1.5), rep(0, p2+p3))
r <- 1
CorM <- 3
TrueR2Cal(alpha,beta,r=3,p0,p1,p2,p3,CorM,LargeSam=F)
TrueR2Cal(alpha,beta,r=3,p0,p1,p2,p3,CorM,LargeSam=T)
}
